requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
Django==1.11.11
django-channels==0.7.0
django-cors-headers==2.2.0
djangorestframework==3.7.7
idna==2.6
lxml==4.2.1
nltk==3.2.5
numpy==1.14.2
oauthlib==2.0.7
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
typing==3.6.4
urllib3==1.22
